# swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Be
  description: |
    The API provides CSR backend API.
basePath: /api/
schemes:
  - https
  - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /equipment/categories/subcategories/{subcategoryId}:
    parameters:
      - name: subcategoryId
        in: path
        required: true
        description: subcategory id
        type: integer
    get:
      summary: Get information about the subcategory of equipment by id.
      security:
        - Bearer: [ ]
      tags:
        - Subcategories
      operationId: GetSubcategoryByID
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SubcategoryResponse"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/SwaggerError"
    delete:
      summary: Delete subcategory of equipment from db by id.
      security:
          - Bearer: [ ]
      tags:
        - Subcategories
      operationId: DeleteSubcategory
      responses:
        200:
          description: subcategory of equipment successfully deleted from db
          schema:
            type: string
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/SwaggerError"
    put:
      summary: Edit equipment subcategory by id.
      security:
        - Bearer: [ ]
      tags:
        - Subcategories
      operationId: UpdateSubcategory
      parameters:
        - name: UpdateSubcategory
          in: body
          required: true
          schema:
            $ref: '#/definitions/NewSubcategory'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SubcategoryResponse"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/SwaggerError"
...
  /pet_size/{petSizeId}:
    parameters:
      - name: petSizeId
        in: path
        required: true
        description: pet size id
        type: integer
    get:
      summary: Get pet size by id
      security:
        - Bearer: [ ]
      tags:
        - PetSize
      operationId: GetPetSize
      responses:
        200:
          description: Pet size has been found
          schema:
            $ref: "#/definitions/PetSizeResponse"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/SwaggerError"
    delete:
      summary: Delete pet size by id
      security:
        - Bearer: [ ]
      tags:
        - PetSize
      operationId: DeletePetSize
      responses:
        200:
          description: Pet size has been deleted
          schema:
            type: string
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/SwaggerError"
    put:
      summary: Edit pet size by id
      security:
        - Bearer: [ ]
      tags:
        - PetSize
      operationId: EditPetSize
      parameters:
        - name: EditPetSize
          in: body
          description: Edit an pet size
          required: true
          schema:
            $ref: "#/definitions/PetSize"
      responses:
        200:
          description: Pet size edited
          schema:
            $ref: "#/definitions/PetSizeResponse"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/SwaggerError"
...
  /pet_kind/{petKindId}:
    parameters:
      - name: petKindId
        in: path
        required: true
        description: pet kind id
        type: integer
    get:
      summary: Get pet kind by id
      security:
        - Bearer: [ ]
      tags:
        - PetKind
      operationId: GetPetKind
      responses:
        200:
          description: Pet kind has been found
          schema:
            $ref: "#/definitions/PetKindResponse"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/SwaggerError"
    delete:
      summary: Delete pet kind by id
      security:
        - Bearer: [ ]
      tags:
        - PetKind
      operationId: DeletePetKind
      responses:
        200:
          description: Pet kind has been deleted
          schema:
            type: string
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/SwaggerError"
    put:
      summary: Edit pet kind by id
      security:
        - Bearer: [ ]
      tags:
        - PetKind
      operationId: EditPetKind
      parameters:
        - name: EditPetKind
          in: body
          description: Edit an pet kind
          required: true
          schema:
            $ref: "#/definitions/PetKind"
      responses:
        200:
          description: Pet kind edited
          schema:
            $ref: "#/definitions/PetKindResponse"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/SwaggerError"
...
definitions:
...
  SwaggerError: # Do not delete it. It is used for generating error responses.
    type: object
    required:
      - code
      - message
    properties:
        code:
            type: integer
            format: int32
        message:
            type: string
        details:
            type: string
  PetKind:
    type: object
    required:
      - name
    properties:
      name:
        type: string
...
  PetKindResponse:
    type: object
    required:
      - id
      - name
    properties:
      id:
        type: integer
      name:
        type: string
...
  Subcategory:
    type: object
    required:
      - id
      - name
      - category
      - max_reservation_time
      - max_reservation_units
    properties:
      id:
        type: integer
      name:
        type: string
      category:
        type: integer
      max_reservation_time:
        type: integer
      max_reservation_units:
        type: integer
...
  SubcategoryResponse:
    type: object
    required:
      - data
    properties:
      data:
        $ref: "#/definitions/Subcategory"
  NewSubcategory:
    type: object
    required:
      - name
      - max_reservation_time
      - max_reservation_units
    properties:
      name:
        type: string
      max_reservation_time:
        type: integer
      max_reservation_units:
        type: integer
  # PetSize
  PetSize:
    type: object
    required:
      - name
      - size
    properties:
      name:
        type: string
        example: "big"
      size:
        type: string
        example: "76x48x53"
      is_universal:
        type: boolean
        example: false
...
  PetSizeResponse:
    type: object
    required:
      - id
      - name
      - size
    properties:
      id:
        type: integer
      name:
        type: string
      size:
        type: string
      is_universal:
        type: boolean

  #Active Areas
...